test(contract): add tests to trigger error message
