+++ packages/contracts/test/UserStateTransitionVerifier.test.ts
@@ -8,12 +8,11 @@ import { Circuit } from '@unirep/circuits'
 import {
     genUserStateTransitionCircuitInput,
     genInputForContract,
-    genNewUserStateTree,
 } from './utils'
 import { Unirep, UserTransitionProof } from '../src'
 import { deployUnirep } from '../src/deploy'
 
-describe('User State Transition', function () {
+describe('User State Transition Verifier', function () {
     this.timeout(600000)
     let accounts: ethers.Signer[]
     let unirepContract: Unirep
@@ -25,214 +24,6 @@ describe('User State Transition', function () {
         accounts = await hardhatEthers.getSigners()
 
         unirepContract = await deployUnirep(<ethers.Wallet>accounts[0])
-
-        unirepContract['userSignUp(uint256)'](
-            user.genIdentityCommitment()
-        ).then((t) => t.wait())
-
-        // UST should be performed after epoch transition
-        // Fast-forward epochLength of seconds
-        const epochLength = (
-            await unirepContract.config()
-        ).epochLength.toNumber()
-        await hardhatEthers.provider.send('evm_increaseTime', [epochLength])
-        unirepContract.beginEpochTransition().then((t) => t.wait())
-    })
-
-    it('Valid user state update inputs should work', async () => {
-        const userStateTree = genNewUserStateTree()
-        const reputationRecords = {}
-        const {
-            startTransitionCircuitInputs,
-            processAttestationCircuitInputs,
-            finalTransitionCircuitInputs,
-        } = await genUserStateTransitionCircuitInput(user, epoch, 0, {
-            userStateTree,
-            reputationRecords,
-        })
-
-        // submit start user state tranisiton proof
-        {
-            const { publicSignals, proof } = await genInputForContract(
-                Circuit.startTransition,
-                startTransitionCircuitInputs
-            )
-
-            unirepContract
-                .startUserStateTransition(publicSignals, proof)
-                .then((t) => t.wait())
-        }
-
-        // submit process attestations proofs
-        {
-            for (const circuitInputs of processAttestationCircuitInputs) {
-                const { publicSignals, proof } = await genInputForContract(
-                    Circuit.processAttestations,
-                    circuitInputs
-                )
-
-                unirepContract
-                    .processAttestations(publicSignals, proof)
-                    .then((t) => t.wait())
-            }
-        }
-
-        // final users state transition proofs
-        const input: UserTransitionProof = await genInputForContract(
-            Circuit.userStateTransition,
-            finalTransitionCircuitInputs
-        )
-        const isValid = await input.verify()
-        expect(isValid, 'Verify user state transition proof off-chain failed')
-            .to.be.true
-        const isProofValid = await unirepContract.verifyUserStateTransition(
-            input.publicSignals,
-            input.proof
-        )
-        expect(isProofValid).to.be.true
-
-        const tx = await unirepContract.updateUserStateRoot(
-            input.publicSignals,
-            input.proof
-        )
-        const receipt = await tx.wait()
-        expect(receipt.status).equal(1)
-
-        const pfIdx = await unirepContract.getProofIndex(input.hash())
-        expect(Number(pfIdx)).not.eq(0)
-
-        for (const nullifier of input.epkNullifiers) {
-            if (!ethers.BigNumber.from(nullifier).eq(0)) {
-                const n = await unirepContract.usedNullifiers(nullifier)
-                expect(
-                    ethers.BigNumber.from(n).eq(0),
-                    'Nullifier is not saved in unirep contract'
-                ).to.be.false
-            }
-        }
-    })
-
-    it('Submit user state transition proof with the same epoch key nullifiers should fail', async () => {
-        const { finalTransitionCircuitInputs: circuitInputs } =
-            await genUserStateTransitionCircuitInput(user, epoch)
-        const input: UserTransitionProof = await genInputForContract(
-            Circuit.userStateTransition,
-            circuitInputs
-        )
-        const isValid = await input.verify()
-        expect(isValid, 'Verify user state transition proof off-chain failed')
-            .to.be.true
-        const isProofValid = await unirepContract.verifyUserStateTransition(
-            input.publicSignals,
-            input.proof
-        )
-        expect(isProofValid).to.be.true
-
-        await expect(
-            unirepContract.updateUserStateRoot(input.publicSignals, input.proof)
-        ).to.be.revertedWithCustomError(unirepContract, 'NullifierAlreadyUsed')
-    })
-
-    it('Invalid GST root should fail', async () => {
-        const user2 = new ZkIdentity()
-        const {
-            startTransitionCircuitInputs,
-            processAttestationCircuitInputs,
-            finalTransitionCircuitInputs: circuitInputs,
-        } = genUserStateTransitionCircuitInput(user2, epoch)
-
-        // submit start user state tranisiton proof
-        {
-            const { publicSignals, proof } = await genInputForContract(
-                Circuit.startTransition,
-                startTransitionCircuitInputs
-            )
-
-            unirepContract
-                .startUserStateTransition(publicSignals, proof)
-                .then((t) => t.wait())
-        }
-
-        // submit process attestations proofs
-        {
-            for (const circuitInputs of processAttestationCircuitInputs) {
-                const { publicSignals, proof } = await genInputForContract(
-                    Circuit.processAttestations,
-                    circuitInputs
-                )
-
-                unirepContract
-                    .processAttestations(publicSignals, proof)
-                    .then((t) => t.wait())
-            }
-        }
-        const input: UserTransitionProof = await genInputForContract(
-            Circuit.userStateTransition,
-            circuitInputs
-        )
-
-        await expect(
-            unirepContract.updateUserStateRoot(input.publicSignals, input.proof)
-        ).to.be.revertedWithCustomError(
-            unirepContract,
-            'InvalidGlobalStateTreeRoot'
-        )
-    })
-
-    it('Invalid user state proof should fail', async () => {
-        const user2 = new ZkIdentity()
-        const {
-            startTransitionCircuitInputs,
-            processAttestationCircuitInputs,
-            finalTransitionCircuitInputs: circuitInputs,
-        } = genUserStateTransitionCircuitInput(user2, epoch)
-
-        // submit start user state tranisiton proof
-        {
-            const { publicSignals, proof } = await genInputForContract(
-                Circuit.startTransition,
-                startTransitionCircuitInputs
-            )
-
-            unirepContract
-                .startUserStateTransition(publicSignals, proof)
-                .then((t) => t.wait())
-        }
-
-        // submit process attestations proofs
-        {
-            for (const circuitInputs of processAttestationCircuitInputs) {
-                const { publicSignals, proof } = await genInputForContract(
-                    Circuit.processAttestations,
-                    circuitInputs
-                )
-
-                unirepContract
-                    .processAttestations(publicSignals, proof)
-                    .then((t) => t.wait())
-            }
-        }
-        const input: UserTransitionProof = await genInputForContract(
-            Circuit.userStateTransition,
-            circuitInputs
-        )
-        const validGSTRoot = await unirepContract.globalStateTree(1)
-        input.publicSignals[input.idx.fromGlobalStateTree] =
-            validGSTRoot.root.toString()
-        const isValid = await input.verify()
-        expect(
-            isValid,
-            'Verify user state transition proof off-chain should fail'
-        ).to.be.false
-        const isProofValid = await unirepContract.verifyUserStateTransition(
-            input.publicSignals,
-            input.proof
-        )
-        expect(isProofValid).to.be.false
-
-        await expect(
-            unirepContract.updateUserStateRoot(input.publicSignals, input.proof)
-        ).to.be.revertedWithCustomError(unirepContract, 'InvalidProof')
     })
 
     it('Proof with wrong epoch should fail', async () => {
